+++ src/pages/RandomList/RandomList.tsx
@@ -11,8 +11,8 @@ import Error from 'pages/Error/Error';
 import MenuCard from 'components/common/MenuCard/MenuCard';
 import AsyncBoundary from 'components/common/AsyncBoundary';
 import Button from 'components/common/Button/Button';
-import BottomSheet from 'components/common/modal/BottomSheet';
-import EndOfListAlertBottomSheet from 'components/common/modal/children/EndOfListAlertBottomSheet';
+import BottomSheet from 'components/common/BottomSheet/BottomSheet';
+import EndOfListAlertBottomSheet from 'components/common/BottomSheet/children/EndOfListAlertBottomSheet';
 import { useRetryMutation } from 'apis/query/useRetryMutation';
 import { useResuggestOneMutation } from 'apis/query/useResuggestOneMutation';
 import { useRandomListMutation } from 'apis/query/useRandomListMutation';
